Rename covid_moonshot -> xchem_infrastructure
