--- covid_moonshot/__init__.py
@@ -1,2 +1,0 @@
-from .core import *
-from .lib import *
