+++ setup.py
@@ -63,5 +63,5 @@ setup(
     # python_requires=">=3.5",          # Python version restrictions
     # Manual control if final package is compressible or not, set False to prevent the .egg from being made
     # zip_safe=False,
-    entry_points={"console_scripts": ["covid-moonshot = covid_moonshot.app:main"]},
+    entry_points={"console_scripts": ["xchem = xchem_infrastructure.app:main"]},
 )
